% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{Arithmetic}
\alias{Arithmetic}
\alias{\%+=\%}
\alias{\%-=\%}
\alias{\%*=\%}
\alias{\%/=\%}
\alias{\%^=\%}
\alias{+}
\title{Arithmetic operators}
\usage{
e1 \%+=\% e2

e1 \%-=\% e2

e1 \%*=\% e2

e1 \%/=\% e2

e1 \%^=\% e2

"+"(x, y)
}
\description{
Basic arithmetic operators that substitute original object with new value.
}
